arbos_revm::api::exec

Type Alias ArbitrumError

pub type ArbitrumError<CTX> = EVMError<<<CTX as ContextTr>::Db as Database>::Error, ArbitrumTransactionError>;

Type alias for the error type of the ArbitrumEvm.

Aliased Type

pub enum ArbitrumError<CTX> {
    Transaction(ArbitrumTransactionError),
    Header(InvalidHeader),
    Database(<<CTX as ContextTr>::Db as Database>::Error),
    Custom(String),
}

Variants

Transaction(ArbitrumTransactionError)

Transaction validation error

Header(InvalidHeader)

Header validation error

Database(<<CTX as ContextTr>::Db as Database>::Error)

Database error

Custom(String)

Custom error

Useful for handler registers where custom logic would want to return their own custom error.

Layout

Note: Unable to compute type layout, possibly due to this type having generic parameters. Layout can only be computed for concrete, fully-instantiated types.